redefine API to provide dictionary metadata. (#197)
* Migrate Smart Browser, Process, and Field.

* Migrate Window and Tabs.
