+++ src/utils/ADempiere/globalMethods.js
@@ -1,2 +1,5 @@
 
-export { isEmptyValue, zeroPad } from '@/utils/ADempiere/valueUtil.js'
+export {
+  isEmptyValue,
+  zeroPad
+} from '@/utils/ADempiere/valueUtils.js'
